refactor(k3): put the forward pass behind a narrow host contract
The executor's step.rs mixed two audiences: the certified model launch
sequence (fixed by the architecture, gated by the golden fixtures,
nobody's daily reading) and the framework-facing facts a reader actually
needs — what crosses the host boundary, why the sequence is capturable,
how a prefill chunk differs from decode.

Move the sequence into executor/forward/ and make the boundary explicit:

* forward/mod.rs — the host contract: 3 staged H2D inputs, 1 D2H sample
  readback, geometry a function of the shape alone (capturable), every
  EP rank launching identically. One screen, the whole story.
* forward/step.rs — the shared certified trunk (embedding, residual
  mixes, MLA, MoE, epilogue), unchanged line for line.
* forward/decode.rs — the decode entry and the per-token KDA walk.
* forward/prefill.rs — the chunk entry, chunkwise FlashKDA, dense-FMHA
  MLA and the boundary sample.
* forward/gemm.rs — moved verbatim.

Pure code motion; no launch, operand or ordering change. Golden decode
gates (12/12) pass on the pruned-224 checkpoint.
