+++ pegainfer-k3/src/executor/forward/step.rs
@@ -1,4 +1,9 @@
-//! One decode step: the certified engine's launch sequence, batched.
+//! One step of the forward pass: the certified engine's launch sequence,
+//! batched. Decode and a prefill chunk run the SAME trunk below; the arms
+//! diverge only at the leaves ([`decode`](super::decode) walks tokens through
+//! the KDA recurrence and absorbed paged MLA, [`prefill`](super::prefill)
+//! runs chunkwise FlashKDA and dense-FMHA MLA) and at the epilogue, which a
+//! chunk skips.
 //!
 //! The order below is the reference `decode_step` line for line. Two families
 //! of call replace a reference spelling rather than reproducing it, and both
@@ -28,7 +33,6 @@ use anyhow::Result;
 use anyhow::ensure;
 use cudarc::driver::CudaSlice;
 use half::bf16;
-use pegainfer_kernels::ops::K3_CONV_WIDTH;
 use pegainfer_kernels::ops::K3_MLA_HEADS;
 use pegainfer_kernels::ops::K3_MOE_QUANT_GROUP;
 use pegainfer_kernels::ops::K3_ROUTER_TOPK;
@@ -40,59 +44,43 @@ use pegainfer_kernels::ops::argmax_bf16_split_into;
 use pegainfer_kernels::ops::copy_hidden_rows_raw_into;
 use pegainfer_kernels::ops::embedding_rows_into;
 use pegainfer_kernels::ops::extract_hidden_rows_raw_into;
-use pegainfer_kernels::ops::gemm_rows_into_checked;
 use pegainfer_kernels::ops::k3_add2_batched_launch;
 use pegainfer_kernels::ops::k3_attnres_mix_batched_launch;
 use pegainfer_kernels::ops::k3_attnres_scores_batched_launch;
-use pegainfer_kernels::ops::k3_conv_silu_batched_launch;
 use pegainfer_kernels::ops::k3_deepgemm_sm100_masked_grouped_fp8_fp4_launch;
-use pegainfer_kernels::ops::k3_flash_kda_fwd_launch;
-use pegainfer_kernels::ops::k3_flash_mla_prefill_fwd_launch;
 use pegainfer_kernels::ops::k3_fp8_scale_pack_ue8m0_launch;
-use pegainfer_kernels::ops::k3_kda_core_batched_launch;
 use pegainfer_kernels::ops::k3_land_batched_launch;
 use pegainfer_kernels::ops::k3_land_rms_norm_rbs_batched_launch;
 use pegainfer_kernels::ops::k3_mega_moe_launch;
 use pegainfer_kernels::ops::k3_mega_write_inputs_launch;
 use pegainfer_kernels::ops::k3_mla_paged_attn_launch;
-use pegainfer_kernels::ops::k3_mla_prefill_expand_k_launch;
-use pegainfer_kernels::ops::k3_mla_prefill_gather_launch;
 use pegainfer_kernels::ops::k3_moe_gather_fp8_quant_masked_launch;
 use pegainfer_kernels::ops::k3_moe_local_route_metadata_launch;
 use pegainfer_kernels::ops::k3_moe_weighted_combine_launch;
 use pegainfer_kernels::ops::k3_mul_sigmoid_batched_launch;
-use pegainfer_kernels::ops::k3_o_norm_gate_batched_launch;
 use pegainfer_kernels::ops::k3_rms_norm_rbs_batched_launch;
 use pegainfer_kernels::ops::k3_router_topk_batched_launch;
 use pegainfer_kernels::ops::k3_situ_and_mul_fp8_quant_masked_launch;
 use pegainfer_kernels::ops::k3_situ_batched_launch;
 use pegainfer_kernels::tensor::DeviceContext;
-use pegainfer_kernels::tensor::DeviceMatrix;
 
-use super::buffers::K3_CONV_STATE;
-use super::buffers::K3_KDA_FUSED;
-use super::buffers::K3_KDA_WSM_PADDED;
-use super::buffers::K3_MLA_FUSED;
-use super::buffers::K3KdaState;
-use super::buffers::K3LayerState;
-use super::buffers::K3Scratch;
-use super::buffers::K3StatePool;
-use super::buffers::copy_rows_2d;
-use super::buffers::parity_pair;
-use super::gemm::K3PartialSpan;
+use super::super::buffers::K3_MLA_FUSED;
+use super::super::buffers::K3LayerState;
+use super::super::buffers::K3Scratch;
+use super::super::buffers::K3StatePool;
+use super::super::buffers::parity_pair;
+use super::super::paged_kv::K3_KV_PAGE_TOKENS;
+use super::super::paged_kv::K3_MLA_LATENT_ROW;
+use super::super::paged_kv::K3PagedKv;
+use super::decode::kda_attention;
 use super::gemm::k3_gemm_full;
-use super::gemm::k3_gemm_partial;
-use super::paged_kv::K3_KV_PAGE_TOKENS;
-use super::paged_kv::K3_MLA_LATENT_ROW;
-use super::paged_kv::K3PagedKv;
+use super::prefill::kda_attention_chunk;
+use super::prefill::mla_attention_chunk_fmha;
 use crate::config::K3_ATTN_INNER;
 use crate::config::K3_DENSE_INTERMEDIATE;
 use crate::config::K3_EXPERT_INTERMEDIATE;
-use crate::config::K3_HEAD_DIM;
-use crate::config::K3_HEADS;
 use crate::config::K3_HIDDEN;
 use crate::config::K3_KV_A_OUT;
-use crate::config::K3_KV_B_OUT;
 use crate::config::K3_KV_LORA_RANK;
 use crate::config::K3_Q_B_OUT;
 use crate::config::K3_Q_LORA_RANK;
@@ -101,7 +89,6 @@ use crate::config::K3_ROUTED_EXPERT_HIDDEN;
 use crate::config::K3_SHARED_INTERMEDIATE;
 use crate::config::K3_VOCAB;
 use crate::model::K3ExpertBankForm;
-use crate::model::K3KdaWeights;
 use crate::model::K3LayerAttention;
 use crate::model::K3LayerMlp;
 use crate::model::K3LayerWeights;
@@ -146,54 +133,7 @@ impl K3StepShape {
     }
 }
 
-/// Advance every row of the bucket by one token and leave the sampled ids in
-/// `scratch.argmax_indices`.
-///
-/// Reads `scratch.token_ids`, `scratch.context_len` and `scratch.kv_row`;
-/// the caller fills those before the step (or before the graph replay).
-///
-/// Every MoE layer issues the same launches in the same order on every rank —
-/// including a step whose batch is empty — which is what lets an
-/// expert-parallel group run without a coordinator.
-pub(crate) fn k3_decode_step(
-    ctx: &DeviceContext,
-    model: &K3RankModel,
-    shape: K3StepShape,
-    state: &mut K3StatePool,
-    scratch: &mut K3Scratch,
-) -> Result<()> {
-    k3_step(ctx, model, shape, false, state, scratch)
-}
-
-/// One prefill chunk: the same batched step, with the bucket's rows carrying
-/// consecutive tokens of ONE sequence instead of independent sequences.
-///
-/// `shape.live_rows` is the chunk's token count and `shape.parity` the KDA
-/// state slab the chunk reads (it lands in the other — parity double-buffers
-/// per chunk here, not per token). The caller stages `token_ids`, ascending
-/// per-token `context_len` and per-token `kv_row` exactly as for decode, and
-/// mirrors the sequence's block-table row across the bucket — causality in the
-/// MLA layers comes from the per-row context length, so the batched attention
-/// kernel serves the chunk unchanged. Only the KDA layers walk the chunk's
-/// tokens through the recurrence sequentially ([`kda_attention_chunk`]);
-/// every other stage runs batched over the rows.
-pub(crate) fn k3_prefill_chunk_step(
-    ctx: &DeviceContext,
-    model: &K3RankModel,
-    shape: K3StepShape,
-    state: &mut K3StatePool,
-    scratch: &mut K3Scratch,
-) -> Result<()> {
-    ensure!(
-        (1..=shape.bucket).contains(&shape.live_rows),
-        "K3 prefill chunk of {} tokens does not fit its {} bucket",
-        shape.live_rows,
-        shape.bucket
-    );
-    k3_step(ctx, model, shape, true, state, scratch)
-}
-
-fn k3_step(
+pub(super) fn k3_step(
     ctx: &DeviceContext,
     model: &K3RankModel,
     shape: K3StepShape,
@@ -390,10 +330,10 @@ fn k3_step(
     }
 
     // A prefill chunk stops here: only the boundary token's sample is ever
-    // read, so the caller runs [`k3_prefill_boundary_sample`] once after the
-    // final chunk instead of paying a chunk-wide lm_head per step — which
-    // also keeps the vocab-wide buffers sized by the decode rows, not the
-    // chunk bucket.
+    // read, so the caller runs [`super::k3_prefill_boundary_sample`] once
+    // after the final chunk instead of paying a chunk-wide lm_head per step —
+    // which also keeps the vocab-wide buffers sized by the decode rows, not
+    // the chunk bucket.
     if prefill_chunk {
         return Ok(());
     }
@@ -444,82 +384,9 @@ fn k3_step(
     )
 }
 
-/// Sample the prefill boundary token: the epilogue the chunk steps skipped,
-/// at `b = 1` over the final chunk's last live row.
-///
-/// The caller must have collapsed the last live row's attention-residual
-/// snapshots to row 0 first (the same collapse the decode handover needs) —
-/// this reads row 0 of `snapshots`. The boundary hidden state is copied from
-/// `hidden[last_row]` into row 0 of the `prefix` scratch, so the sample lands
-/// in `argmax_indices[0]`.
-pub(crate) fn k3_prefill_boundary_sample(
-    ctx: &DeviceContext,
-    model: &K3RankModel,
-    last_row: usize,
-    snapshots: &CudaSlice<bf16>,
-    scratch: &mut K3Scratch,
-) -> Result<()> {
-    copy_rows_2d(
-        ctx,
-        &scratch.hidden,
-        last_row * K3_HIDDEN,
-        K3_HIDDEN,
-        &mut scratch.prefix,
-        0,
-        K3_HIDDEN,
-        1,
-        K3_HIDDEN,
-    )?;
-    attn_res(
-        ctx,
-        1,
-        model.blocks,
-        &scratch.prefix,
-        snapshots,
-        &model.sw_out,
-        &mut scratch.scores,
-        &mut scratch.mixed,
-    )?;
-    k3_rms_norm_rbs_batched_launch(
-        ctx,
-        1,
-        K3_HIDDEN,
-        &scratch.mixed,
-        &model.gamma_final.data,
-        &mut scratch.normed,
-    )?;
-    k3_gemm_full(
-        ctx,
-        &model.w_lm,
-        &scratch.normed,
-        1,
-        &mut scratch.logit_partial,
-    )?;
-    k3_land_batched_launch(
-        ctx,
-        1,
-        K3_VOCAB,
-        K3_VOCAB,
-        0,
-        1,
-        &scratch.logit_partial,
-        &mut scratch.logits,
-    )?;
-    argmax_bf16_split_into(
-        ctx,
-        &scratch.logits,
-        1,
-        K3_VOCAB,
-        &mut scratch.argmax_partial_values,
-        &mut scratch.argmax_partial_indices,
-        &mut scratch.argmax_values,
-        &mut scratch.argmax_indices,
-    )
-}
-
 /// Score the `blocks + 1` attention-residual candidates and mix them.
 #[allow(clippy::too_many_arguments)]
-fn attn_res(
+pub(super) fn attn_res(
     ctx: &DeviceContext,
     b: usize,
     blocks: usize,
@@ -546,492 +413,6 @@ fn copy_rows(
     ctx.stream
         .memcpy_dtod(&origin, &mut destination)
         .map_err(|error| anyhow::anyhow!("K3 residual copy failed: {error}"))
-}
-
-// ── KDA ─────────────────────────────────────────────────────────────────
-
-#[allow(clippy::too_many_arguments)]
-fn kda_attention(
-    ctx: &DeviceContext,
-    b: usize,
-    layer: &K3LayerWeights,
-    w: &K3KdaWeights,
-    recurrent_read: &CudaSlice<f32>,
-    recurrent_write: &mut CudaSlice<f32>,
-    conv_read: &[CudaSlice<bf16>; 3],
-    conv_write: &mut [CudaSlice<bf16>; 3],
-    s: &mut K3Scratch,
-) -> Result<()> {
-    k3_rms_norm_rbs_batched_launch(
-        ctx,
-        b,
-        K3_HIDDEN,
-        &s.mixed,
-        &layer.gamma_in.data,
-        &mut s.normed,
-    )?;
-
-    // The output-gate quarter of the fused q|k|v|gate projection, placed where
-    // the certified `(4 * inner, inner, 3 * inner)` landing span expects it.
-    k3_gemm_partial(
-        ctx,
-        &w.wbig,
-        3 * K3_ATTN_INNER,
-        K3_ATTN_INNER,
-        &s.normed,
-        b,
-        &mut s.kda_gate_partial,
-        K3PartialSpan {
-            offset: 3 * K3_ATTN_INNER,
-            stride: K3_KDA_FUSED,
-        },
-    )?;
-    k3_gemm_full(ctx, &w.wsm, &s.normed, b, &mut s.kda_wsm_partial)?;
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_KDA_WSM_PADDED,
-        K3_HEADS,
-        0,
-        1,
-        &s.kda_wsm_partial,
-        &mut s.beta,
-    )?;
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_KDA_WSM_PADDED,
-        K3_HEAD_DIM,
-        K3_HEADS,
-        1,
-        &s.kda_wsm_partial,
-        &mut s.forget_low,
-    )?;
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_KDA_FUSED,
-        K3_ATTN_INNER,
-        3 * K3_ATTN_INNER,
-        1,
-        &s.kda_gate_partial,
-        &mut s.out_gate,
-    )?;
-    k3_gemm_full(ctx, &w.w_f_b, &s.forget_low, b, &mut s.kda_forget_partial)?;
-
-    kda_conv_stream(
-        ctx,
-        b,
-        &w.wbig,
-        0,
-        &w.cw_q,
-        &conv_read[0],
-        &mut conv_write[0],
-        &s.normed,
-        &mut s.kda_conv_partial,
-        &mut s.conv_x,
-        &mut s.conv_q,
-    )?;
-    kda_conv_stream(
-        ctx,
-        b,
-        &w.wbig,
-        K3_ATTN_INNER,
-        &w.cw_k,
-        &conv_read[1],
-        &mut conv_write[1],
-        &s.normed,
-        &mut s.kda_conv_partial,
-        &mut s.conv_x,
-        &mut s.conv_k,
-    )?;
-    kda_conv_stream(
-        ctx,
-        b,
-        &w.wbig,
-        2 * K3_ATTN_INNER,
-        &w.cw_v,
-        &conv_read[2],
-        &mut conv_write[2],
-        &s.normed,
-        &mut s.kda_conv_partial,
-        &mut s.conv_x,
-        &mut s.conv_v,
-    )?;
-
-    k3_kda_core_batched_launch(
-        ctx,
-        b,
-        K3_HEADS,
-        K3_HEAD_DIM,
-        1,
-        &s.conv_q,
-        &s.conv_k,
-        &s.conv_v,
-        &s.kda_forget_partial,
-        &w.dt_bias,
-        &w.a_log,
-        &s.beta,
-        &s.out_gate,
-        &w.gamma_o,
-        recurrent_read,
-        recurrent_write,
-        &mut s.gated,
-    )?;
-    k3_gemm_full(ctx, &w.w_o, &s.gated, b, &mut s.hidden_partial)?;
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_HIDDEN,
-        K3_HIDDEN,
-        0,
-        1,
-        &s.hidden_partial,
-        &mut s.attn_out,
-    )
-}
-
-/// One q/k/v stream: its band of the fused projection, then the window.
-///
-/// The band goes to a partial of its own rather than into the fused one: the
-/// convolution reads a dense `[rows, inner]` partial, and slicing a column
-/// block out of the fused product would cost a strided f32 copy per stream.
-#[allow(clippy::too_many_arguments)]
-fn kda_conv_stream(
-    ctx: &DeviceContext,
-    b: usize,
-    fused: &DeviceMatrix,
-    band: usize,
-    taps: &CudaSlice<f32>,
-    window_read: &CudaSlice<bf16>,
-    window_write: &mut CudaSlice<bf16>,
-    normed: &CudaSlice<bf16>,
-    partial: &mut CudaSlice<f32>,
-    landed: &mut CudaSlice<bf16>,
-    out: &mut CudaSlice<bf16>,
-) -> Result<()> {
-    k3_gemm_partial(
-        ctx,
-        fused,
-        band,
-        K3_ATTN_INNER,
-        normed,
-        b,
-        partial,
-        K3PartialSpan::whole(K3_ATTN_INNER),
-    )?;
-    k3_conv_silu_batched_launch(
-        ctx,
-        b,
-        K3_ATTN_INNER,
-        K3_CONV_WIDTH,
-        1,
-        partial,
-        taps,
-        window_read,
-        landed,
-        out,
-        window_write,
-    )
-}
-
-/// The KDA layer of a prefill chunk: batched projections, a batched
-/// convolution over prebuilt windows, and the delta rule as one chunkwise
-/// FlashKDA forward (third_party/flash-kda, MIT, MoonshotAI) — the per-token
-/// walk collapses into two launches per layer.
-///
-/// Numerics: the convolution windows are the landed bf16 inputs themselves
-/// (`window[t][j] = x[t - 3 + j]`, carried window for the first rows), so the
-/// conv is bitwise against sequential stepping. FlashKDA computes the same
-/// delta rule the fused TileLang core spells — same gate formula from the
-/// same pre-activation landing, beta sigmoid in-kernel — but with an f32 q/k
-/// l2norm chain (the TileLang core mirrors the reference's bf16 chain) and
-/// chunkwise accumulation order, and the projections run at the chunk's
-/// bucket. Chunked prefill is therefore held to the fixture's noise floor,
-/// like any cross-bucket comparison; decode keeps the bit-matched fused core.
-fn kda_attention_chunk(
-    ctx: &DeviceContext,
-    shape: K3StepShape,
-    layer: &K3LayerWeights,
-    w: &K3KdaWeights,
-    kda_state: &mut K3KdaState,
-    s: &mut K3Scratch,
-) -> Result<()> {
-    let b = shape.bucket;
-    let tokens = shape.live_rows;
-    let start_parity = shape.parity;
-    // Batched projections — the decode arm's launches, at the chunk's bucket.
-    k3_rms_norm_rbs_batched_launch(
-        ctx,
-        b,
-        K3_HIDDEN,
-        &s.mixed,
-        &layer.gamma_in.data,
-        &mut s.normed,
-    )?;
-    k3_gemm_partial(
-        ctx,
-        &w.wbig,
-        3 * K3_ATTN_INNER,
-        K3_ATTN_INNER,
-        &s.normed,
-        b,
-        &mut s.kda_gate_partial,
-        K3PartialSpan {
-            offset: 3 * K3_ATTN_INNER,
-            stride: K3_KDA_FUSED,
-        },
-    )?;
-    k3_gemm_full(ctx, &w.wsm, &s.normed, b, &mut s.kda_wsm_partial)?;
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_KDA_WSM_PADDED,
-        K3_HEADS,
-        0,
-        1,
-        &s.kda_wsm_partial,
-        &mut s.beta,
-    )?;
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_KDA_WSM_PADDED,
-        K3_HEAD_DIM,
-        K3_HEADS,
-        1,
-        &s.kda_wsm_partial,
-        &mut s.forget_low,
-    )?;
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_KDA_FUSED,
-        K3_ATTN_INNER,
-        3 * K3_ATTN_INNER,
-        1,
-        &s.kda_gate_partial,
-        &mut s.out_gate,
-    )?;
-    k3_gemm_full(ctx, &w.w_f_b, &s.forget_low, b, &mut s.kda_forget_partial)?;
-
-    kda_conv_stream_chunk(
-        ctx,
-        b,
-        tokens,
-        start_parity,
-        &w.wbig,
-        0,
-        &w.cw_q,
-        &mut kda_state.conv,
-        0,
-        &s.normed,
-        &mut s.kda_conv_partial,
-        &mut s.conv_x,
-        &mut s.conv_window,
-        &mut s.conv_window_next,
-        &mut s.conv_q,
-    )?;
-    kda_conv_stream_chunk(
-        ctx,
-        b,
-        tokens,
-        start_parity,
-        &w.wbig,
-        K3_ATTN_INNER,
-        &w.cw_k,
-        &mut kda_state.conv,
-        1,
-        &s.normed,
-        &mut s.kda_conv_partial,
-        &mut s.conv_x,
-        &mut s.conv_window,
-        &mut s.conv_window_next,
-        &mut s.conv_k,
-    )?;
-    kda_conv_stream_chunk(
-        ctx,
-        b,
-        tokens,
-        start_parity,
-        &w.wbig,
-        2 * K3_ATTN_INNER,
-        &w.cw_v,
-        &mut kda_state.conv,
-        2,
-        &s.normed,
-        &mut s.kda_conv_partial,
-        &mut s.conv_x,
-        &mut s.conv_window,
-        &mut s.conv_window_next,
-        &mut s.conv_v,
-    )?;
-
-    // The chunkwise delta rule: the whole chunk through the vendored FlashKDA
-    // forward (third_party/flash-kda) — two launches instead of a per-token
-    // walk. The pre-activation gate lands bf16 first, the fused core's own
-    // `bf16(Σ gp)` landing; FlashKDA adds dt_bias and applies the activation
-    // in-kernel, from the same formula. The recurrent state reads the
-    // start-parity slab and lands in the other one: under the chunkwise
-    // kernel, parity is a per-chunk double buffer, not a per-token flip.
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_ATTN_INNER,
-        K3_ATTN_INNER,
-        0,
-        1,
-        &s.kda_forget_partial,
-        &mut s.kda_g,
-    )?;
-    {
-        let (recurrent_read, recurrent_write) = parity_pair(&mut kda_state.recurrent, start_parity);
-        k3_flash_kda_fwd_launch(
-            ctx,
-            tokens,
-            K3_HEADS,
-            &s.conv_q,
-            &s.conv_k,
-            &s.conv_v,
-            &s.kda_g,
-            &s.beta,
-            &mut s.kda_beta_t,
-            &w.a_log,
-            &w.dt_bias,
-            recurrent_read,
-            recurrent_write,
-            &mut s.kda_attn,
-            &mut s.flash_kda_ws,
-            (K3_HEAD_DIM as f32).powf(-0.5),
-            crate::config::K3_KDA_GATE_LOWER_BOUND as f32,
-        )?;
-    }
-    // o_norm × output gate — the fused core's tail as its own batched launch;
-    // rows past the chunk carry stale data the step discards with the rest of
-    // the padding.
-    k3_o_norm_gate_batched_launch(
-        ctx,
-        b,
-        K3_HEADS,
-        K3_HEAD_DIM,
-        &s.kda_attn,
-        &s.out_gate,
-        &w.gamma_o,
-        &mut s.gated,
-    )?;
-
-    // Batched output projection; rows past the chunk carry stale data the
-    // step discards with the rest of the padding.
-    k3_gemm_full(ctx, &w.w_o, &s.gated, b, &mut s.hidden_partial)?;
-    k3_land_batched_launch(
-        ctx,
-        b,
-        K3_HIDDEN,
-        K3_HIDDEN,
-        0,
-        1,
-        &s.hidden_partial,
-        &mut s.attn_out,
-    )
-}
-
-/// One q/k/v stream of a prefill chunk: the batched band projection, the
-/// window build, one batched convolution, and the carry into the next chunk.
-#[allow(clippy::too_many_arguments)]
-fn kda_conv_stream_chunk(
-    ctx: &DeviceContext,
-    b: usize,
-    tokens: usize,
-    start_parity: usize,
-    fused: &DeviceMatrix,
-    band: usize,
-    taps: &CudaSlice<f32>,
-    conv_state: &mut [[CudaSlice<bf16>; 3]; 2],
-    stream_index: usize,
-    normed: &CudaSlice<bf16>,
-    partial: &mut CudaSlice<f32>,
-    xs: &mut CudaSlice<bf16>,
-    window: &mut CudaSlice<bf16>,
-    window_next: &mut CudaSlice<bf16>,
-    out: &mut CudaSlice<bf16>,
-) -> Result<()> {
-    let inner = K3_ATTN_INNER;
-    k3_gemm_partial(
-        ctx,
-        fused,
-        band,
-        inner,
-        normed,
-        b,
-        partial,
-        K3PartialSpan::whole(inner),
-    )?;
-    // Land the chunk's inputs once: the window entries ARE these bf16 rows,
-    // the same cast the conv kernel itself applies.
-    k3_land_batched_launch(ctx, b, inner, inner, 0, 1, partial, xs)?;
-    // Row t's window slot j holds input `t - K3_CONV_STATE + j`: from the
-    // chunk itself once that token exists, from the carried window before it.
-    {
-        let carry = &conv_state[start_parity][stream_index];
-        for j in 0..K3_CONV_STATE {
-            let lead = K3_CONV_STATE - j;
-            if tokens > lead {
-                copy_rows_2d(
-                    ctx,
-                    xs,
-                    0,
-                    inner,
-                    window,
-                    (lead * K3_CONV_STATE + j) * inner,
-                    K3_CONV_STATE * inner,
-                    tokens - lead,
-                    inner,
-                )?;
-            }
-            for t in 0..lead.min(tokens) {
-                copy_rows_2d(
-                    ctx,
-                    carry,
-                    (t + j) * inner,
-                    inner,
-                    window,
-                    (t * K3_CONV_STATE + j) * inner,
-                    inner,
-                    1,
-                    inner,
-                )?;
-            }
-        }
-    }
-    k3_conv_silu_batched_launch(
-        ctx,
-        b,
-        inner,
-        K3_CONV_WIDTH,
-        1,
-        partial,
-        taps,
-        window,
-        xs,
-        out,
-        window_next,
-    )?;
-    // The carry into the next chunk is the successor window of the chunk's
-    // last token — for a chunk shorter than the window it already folds the
-    // slots carried in above. It lands in the other parity slab, agreeing
-    // with the recurrent state's per-chunk double buffering.
-    let end_parity = start_parity ^ 1;
-    copy_rows_2d(
-        ctx,
-        window_next,
-        (tokens - 1) * K3_CONV_STATE * inner,
-        K3_CONV_STATE * inner,
-        &mut conv_state[end_parity][stream_index],
-        0,
-        K3_CONV_STATE * inner,
-        1,
-        K3_CONV_STATE * inner,
-    )
 }
 
 // ── MLA ─────────────────────────────────────────────────────────────────
@@ -1175,73 +556,6 @@ fn mla_attention(
         1,
         &s.hidden_partial,
         &mut s.attn_out,
-    )
-}
-
-/// One prefill chunk's MLA attention over FlashMLA's dense FMHA.
-///
-/// The workspace covers the whole context (`max_ctx` rows), so the vLLM
-/// context loop degenerates to a single call: `t_q = live_rows` queries from
-/// `s.query` attend `t_kv = chunk_start + live_rows` keys expanded from the
-/// gathered latent, and `CausalMask<false>`'s Q-at-the-end alignment gives
-/// chunk token `i` exactly `chunk_start + i + 1` visible keys. Prefill drives
-/// slot 0, whose block-table row leads `table_dev`.
-fn mla_attention_chunk_fmha(
-    ctx: &DeviceContext,
-    shape: K3StepShape,
-    w: &K3MlaWeights,
-    kv: &K3PagedKv,
-    mla_index: usize,
-    s: &mut K3Scratch,
-) -> Result<()> {
-    let t_q = shape.live_rows;
-    let t_kv = shape.chunk_start + t_q;
-    ensure!(
-        t_kv * K3_KV_LORA_RANK <= s.mla_ctx_latent.data.len(),
-        "K3 MLA prefill workspace of {} tokens cannot span the {t_kv}-token context",
-        s.mla_ctx_latent.data.len() / K3_KV_LORA_RANK
-    );
-    k3_mla_prefill_gather_launch(
-        ctx,
-        t_kv,
-        &kv.slab,
-        &kv.table_dev,
-        kv.page_stride(),
-        mla_index * K3_KV_PAGE_TOKENS * K3_MLA_LATENT_ROW,
-        &mut s.mla_ctx_latent.data,
-        &mut s.mla_ctx_rope,
-    )?;
-    s.mla_ctx_latent.seq_len = t_kv;
-    s.mla_ctx_nope_v.seq_len = t_kv;
-    gemm_rows_into_checked(
-        ctx,
-        &w.w_kv_b,
-        0,
-        K3_KV_B_OUT,
-        &s.mla_ctx_latent,
-        &mut s.mla_ctx_nope_v,
-    )?;
-    k3_mla_prefill_expand_k_launch(
-        ctx,
-        t_kv,
-        K3_MLA_HEADS,
-        &s.mla_ctx_nope_v.data,
-        &s.mla_ctx_rope,
-        &mut s.mla_ctx_k,
-    )?;
-    // The decode kernel reads the softmax scale as a bf16 device scalar; feed
-    // the FMHA the same rounded constant so the two paths agree on it.
-    let scale = bf16::from_f64(crate::model::k3_mla_scale()).to_f32();
-    k3_flash_mla_prefill_fwd_launch(
-        ctx,
-        t_q,
-        t_kv,
-        K3_MLA_HEADS,
-        &s.query,
-        &s.mla_ctx_k,
-        &s.mla_ctx_nope_v.data,
-        &mut s.attn,
-        scale,
     )
 }
